+++ flash-core/src/main/java/cn/enilu/flash/dao/business/BidRepository.java
@@ -42,7 +42,7 @@ public interface BidRepository extends PagingAndSortingRepository<Bid, Long>, Jp
     @Transactional
     @Modifying
     @Query(nativeQuery = true, value = "update t_biz_bid set status = status+1, confirmed_quantity=?2, confirmed_price=?3, confirmed_quantity_unit=?4, confirmed_price_unit=?5 where id=?1")
-    void moveToNextStatusWithQuantityPrice(Long id, Integer confirmedQuantity, Float confirmedPrice, String confirmedQuantityUnit, String confirmedPriceUnit);
+    void moveToNextStatusWithQuantityPrice(Long id, Double confirmedQuantity, Double confirmedPrice, String confirmedQuantityUnit, String confirmedPriceUnit);
 
     @Transactional
     @Modifying
@@ -67,5 +67,5 @@ public interface BidRepository extends PagingAndSortingRepository<Bid, Long>, Jp
     @Transactional
     @Modifying
     @Query(nativeQuery = true, value = "update t_biz_bid set quantity=?2, unit_of_bid_quantity=?3, price=?4, unit_of_bid_price=?5 where id=?1")
-    void updateQuantityAndPrice(Long id, Integer quantity, String unitOfBidQuantity, Float price, String unitOfBidPrice);
+    void updateQuantityAndPrice(Long id, Double quantity, String unitOfBidQuantity, Double price, String unitOfBidPrice);
 }
